Fix checks message expansion issue with undefined reference
We are sure the reference label will eventually be found, but if the label
comes after it, the missing default value in \zref@extract leads to an empty
message argument, which then consumes the next token and results in an
incomplete \if, and then an error, which became persistent, since it prevented
the reference from being defined.  The 'unknown' value solves this, and will
eventually get replaced by the right one as the compilation settles.